+++ src/Libraries/Microsoft.Extensions.AsyncState/IAsyncContext.cs
@@ -8,7 +8,7 @@ namespace Microsoft.Extensions.AsyncState;
 
 /// <summary>
 /// Provides access to the current async context.
-/// Some implementations of this interface may not be thread safe.
+/// Some implementations of this interface might not be thread safe.
 /// </summary>
 /// <typeparam name="T">The type of the asynchronous state.</typeparam>
 public interface IAsyncContext<T>